misc: Move auth constants to separate file
This simplifies avoiding circular imports when trying to use auth
handlers.
