+++ backend/endpoints/tests/test_oauth.py
@@ -2,7 +2,7 @@ import pytest
 from endpoints.auth import ACCESS_TOKEN_EXPIRE_MINUTES
 from fastapi.exceptions import HTTPException
 from fastapi.testclient import TestClient
-from handler.auth.base_handler import WRITE_SCOPES
+from handler.auth.constants import WRITE_SCOPES
 from main import app
