+++ backend/handler/auth/base_handler.py
@@ -1,71 +1,17 @@
-import enum
 import uuid
 from datetime import datetime, timedelta, timezone
-from typing import Any, Final
+from typing import Any
 
 from config import OIDC_ENABLED, ROMM_AUTH_SECRET_KEY
 from exceptions.auth_exceptions import OAuthCredentialsException, UserDisabledException
 from fastapi import HTTPException, status
+from handler.auth.constants import ALGORITHM, DEFAULT_OAUTH_TOKEN_EXPIRY
 from joserfc import jwt
 from joserfc.errors import BadSignatureError
 from joserfc.jwk import OctKey
 from logger.logger import log
 from passlib.context import CryptContext
 from starlette.requests import HTTPConnection
-
-ALGORITHM: Final = "HS256"
-DEFAULT_OAUTH_TOKEN_EXPIRY: Final = timedelta(minutes=15)
-
-
-class Scope(enum.StrEnum):
-    ME_READ = "me.read"
-    ME_WRITE = "me.write"
-    ROMS_READ = "roms.read"
-    ROMS_WRITE = "roms.write"
-    ROMS_USER_READ = "roms.user.read"
-    ROMS_USER_WRITE = "roms.user.write"
-    PLATFORMS_READ = "platforms.read"
-    PLATFORMS_WRITE = "platforms.write"
-    ASSETS_READ = "assets.read"
-    ASSETS_WRITE = "assets.write"
-    FIRMWARE_READ = "firmware.read"
-    FIRMWARE_WRITE = "firmware.write"
-    COLLECTIONS_READ = "collections.read"
-    COLLECTIONS_WRITE = "collections.write"
-    USERS_READ = "users.read"
-    USERS_WRITE = "users.write"
-    TASKS_RUN = "tasks.run"
-
-
-DEFAULT_SCOPES_MAP: Final = {
-    Scope.ME_READ: "View your profile",
-    Scope.ME_WRITE: "Modify your profile",
-    Scope.ROMS_READ: "View ROMs",
-    Scope.PLATFORMS_READ: "View platforms",
-    Scope.ASSETS_READ: "View assets",
-    Scope.ASSETS_WRITE: "Modify assets",
-    Scope.FIRMWARE_READ: "View firmware",
-    Scope.ROMS_USER_READ: "View user-rom properties",
-    Scope.ROMS_USER_WRITE: "Modify user-rom properties",
-    Scope.COLLECTIONS_READ: "View collections",
-    Scope.COLLECTIONS_WRITE: "Modify collections",
-}
-
-WRITE_SCOPES_MAP: Final = {
-    Scope.ROMS_WRITE: "Modify ROMs",
-    Scope.PLATFORMS_WRITE: "Modify platforms",
-    Scope.FIRMWARE_WRITE: "Modify firmware",
-}
-
-FULL_SCOPES_MAP: Final = {
-    Scope.USERS_READ: "View users",
-    Scope.USERS_WRITE: "Modify users",
-    Scope.TASKS_RUN: "Run tasks",
-}
-
-DEFAULT_SCOPES: Final = list(DEFAULT_SCOPES_MAP.keys())
-WRITE_SCOPES: Final = DEFAULT_SCOPES + list(WRITE_SCOPES_MAP.keys())
-FULL_SCOPES: Final = WRITE_SCOPES + list(FULL_SCOPES_MAP.keys())
 
 
 class AuthHandler:
