+++ backend/handler/auth/tests/test_auth.py
@@ -3,7 +3,7 @@ from base64 import b64encode
 import pytest
 from fastapi.exceptions import HTTPException
 from handler.auth import auth_handler, oauth_handler
-from handler.auth.base_handler import WRITE_SCOPES
+from handler.auth.constants import WRITE_SCOPES
 from handler.auth.hybrid_auth import HybridAuthBackend
 from handler.database import db_user_handler
 from models.user import User
