+++ backend/models/user.py
@@ -4,13 +4,13 @@ import enum
 from datetime import datetime, timezone
 from typing import TYPE_CHECKING
 
+from handler.auth.constants import DEFAULT_SCOPES, FULL_SCOPES, WRITE_SCOPES, Scope
 from models.base import BaseModel
 from sqlalchemy import DateTime, Enum, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from starlette.authentication import SimpleUser
 
 if TYPE_CHECKING:
-    from handler.auth.base_handler import Scope
     from models.assets import Save, Screenshot, State
     from models.collection import Collection
     from models.rom import RomUser
@@ -47,8 +47,6 @@ class User(BaseModel, SimpleUser):
 
     @property
     def oauth_scopes(self) -> list[Scope]:
-        from handler.auth.base_handler import DEFAULT_SCOPES, FULL_SCOPES, WRITE_SCOPES
-
         if self.role == Role.ADMIN:
             return FULL_SCOPES
